requirements.txt
Flask==3.0.0
pymongo==4.6.0
bcrypt==4.1.2
google-generativeai==0.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
reportlab==4.0.7
Pillow==10.1.0
pytz==2023.3
